Simple fixes accumulated after 1.3.0. busybox-1.3.0.ash.patch busybox-1.3.0.bb_strtou.patch busybox-1.3.0.CONFIG_FEATURE_TAR_GNU_EXTENSIONS.patch busybox-1.3.0.dhcprelay.patch busybox-1.3.0.dpkg_ar.patch busybox-1.3.0.find.patch busybox-1.3.0.mount.patch busybox-1.3.0.perror.patch busybox-1.3.0.sed.patch busybox-1.3.0.shadow.patch busybox-1.3.0.xregcomp.patch

+++ include/xatonum.h
@@ -115,7 +115,7 @@ extern inline
 unsigned long bb_strtoul(const char *arg, char **endp, int base)
 { return bb_strtoull(arg, endp, base); }
 extern inline
-unsigned long bb_strtol(const char *arg, char **endp, int base)
+long bb_strtol(const char *arg, char **endp, int base)
 { return bb_strtoll(arg, endp, base); }
 #else
 unsigned long bb_strtoul(const char *arg, char **endp, int base);
@@ -124,21 +124,21 @@ long bb_strtol(const char *arg, char **endp, int base);
 
 #if UINT_MAX == ULLONG_MAX
 extern inline
-unsigned long bb_strtou(const char *arg, char **endp, int base)
+unsigned bb_strtou(const char *arg, char **endp, int base)
 { return bb_strtoull(arg, endp, base); }
 extern inline
-unsigned long bb_strtoi(const char *arg, char **endp, int base)
+int bb_strtoi(const char *arg, char **endp, int base)
 { return bb_strtoll(arg, endp, base); }
 #elif UINT_MAX == ULONG_MAX
 extern inline
-unsigned long bb_strtou(const char *arg, char **endp, int base)
+unsigned bb_strtou(const char *arg, char **endp, int base)
 { return bb_strtoul(arg, endp, base); }
 extern inline
-unsigned long bb_strtoi(const char *arg, char **endp, int base)
+int bb_strtoi(const char *arg, char **endp, int base)
 { return bb_strtol(arg, endp, base); }
 #else
-unsigned long bb_strtou(const char *arg, char **endp, int base);
-long bb_strtoi(const char *arg, char **endp, int base);
+unsigned bb_strtou(const char *arg, char **endp, int base);
+int bb_strtoi(const char *arg, char **endp, int base);
 #endif
 
 int BUG_bb_strtou32_unimplemented(void);
